Add support for copanion object
