+++ macros/src/main/scala-2/com/softwaremill/macwire/internals/ConstructorCrimper.scala
@@ -116,60 +116,26 @@ object ConstructorCrimper {
   def constructorTree[C <: blackbox.Context](
       c: C,
       log: Logger
-  )(targetType: c.Type, resolver: (c.Symbol, c.Type) => c.Tree): Option[c.Tree] = {
-    import c.universe._
+  )(targetType: c.Type, resolver: (c.Symbol, c.Type) => c.Tree): Option[c.Tree] =
+    constructorFactory(c, log)(targetType).map { case (paramLists, factory) =>
+      import c.universe._
 
-    lazy val targetTypeD: Type = targetType.dealias
+      lazy val targetTypeD: Type = targetType.dealias
 
-    lazy val constructor: Option[Symbol] = ConstructorCrimper.constructor(c, log)(targetType)
+      def wireConstructorParams(paramLists: List[List[Symbol]]): List[List[Tree]] =
+        paramLists.map(_.map(p => resolver(p, /*SI-4751*/ paramType(c)(targetTypeD, p))))
 
-    lazy val constructorParamLists: Option[List[List[Symbol]]] =
-      constructor.map(_.asMethod.paramLists.filterNot(_.headOption.exists(_.isImplicit)))
+      def constructorArgs: List[List[Tree]] = log.withBlock("Looking for targetConstructor arguments") {
+        wireConstructorParams(paramLists)
+      }
 
-    def constructorArgs: Option[List[List[Tree]]] = log.withBlock("Looking for targetConstructor arguments") {
-      constructorParamLists.map(wireConstructorParams(_))
+      factory(constructorArgs)
     }
 
-    def wireConstructorParams(paramLists: List[List[Symbol]]): List[List[Tree]] =
-      paramLists.map(_.map(p => resolver(p, /*SI-4751*/ paramType(c)(targetTypeD, p))))
-
-    log.withBlock(s"Creating Constructor Tree for $targetType") {
-      val constructionMethodTree: Tree = Select(New(Ident(targetTypeD.typeSymbol)), termNames.CONSTRUCTOR)
-      constructorArgs.map(_.foldLeft(constructionMethodTree)((acc: Tree, args: List[Tree]) => Apply(acc, args)))
-    }
-  }
-
-  def constructorTreeV2[C <: blackbox.Context](
+  def constructorFactory[C <: blackbox.Context](
       c: C,
       log: Logger
-  )(targetType: c.Type, resolver: (c.Symbol, c.Type) => Option[c.Tree]): Option[c.Tree] = {
-    import c.universe._
-
-    lazy val targetTypeD: Type = targetType.dealias
-
-    lazy val constructor: Option[Symbol] = ConstructorCrimper.constructor(c, log)(targetType)
-
-    lazy val constructorParamLists: Option[List[List[Symbol]]] =
-      constructor.map(_.asMethod.paramLists.filterNot(_.headOption.exists(_.isImplicit)))
-
-    def constructorArgs: Option[List[List[Option[Tree]]]] = log.withBlock("Looking for targetConstructor arguments") {
-      constructorParamLists.map(wireConstructorParams(_))
-    }
-
-    def wireConstructorParams(paramLists: List[List[Symbol]]): List[List[Option[Tree]]] =
-      paramLists.map(_.map(p => resolver(p, /*SI-4751*/ paramType(c)(targetTypeD, p))))
-
-    log.withBlock(s"Creating Constructor Tree for $targetType") {
-      val constructionMethodTree: Tree = Select(New(Ident(targetTypeD.typeSymbol)), termNames.CONSTRUCTOR)
-      val flattenConstructorArgs = constructorArgs.flatMap(l => sequence(l.map(sequence)))//FIXME it breaks the current macwire's philosophy....
-      flattenConstructorArgs.map(_.foldLeft(constructionMethodTree)((acc: Tree, args: List[Tree]) => Apply(acc, args)))
-    }
-  }
-
-  def constructorV3[C <: blackbox.Context](
-    c: C,
-    log: Logger
-  )(targetType: c.Type): Option[(List[List[c.universe.Symbol]], List[List[c.Tree]] => c.Tree)] = {
+  )(targetType: c.Type): Option[(List[List[c.Symbol]], List[List[c.Tree]] => c.Tree)] = {
     import c.universe._
 
     lazy val targetTypeD: Type = targetType.dealias
@@ -179,12 +145,12 @@ object ConstructorCrimper {
     val constructorParamLists: Option[List[List[Symbol]]] =
       constructor.map(_.asMethod.paramLists.filterNot(_.headOption.exists(_.isImplicit)))
 
-    def fff(constructorArgs: List[List[Tree]]): Tree = {
-        val constructionMethodTree: Tree = Select(New(Ident(targetTypeD.typeSymbol)), termNames.CONSTRUCTOR)
+    def factory(constructorArgs: List[List[Tree]]): Tree = {
+      val constructionMethodTree: Tree = Select(New(Ident(targetTypeD.typeSymbol)), termNames.CONSTRUCTOR)
       constructorArgs.foldLeft(constructionMethodTree)((acc: Tree, args: List[Tree]) => Apply(acc, args))
-  }
+    }
 
-    constructorParamLists.map(cpl => (cpl, fff(_)))
+    constructorParamLists.map(cpl => (cpl, factory(_)))
   }
 
 }
