+++ macrosAutoCats/src/main/scala/com/softwaremill/macwire/autocats/package.scala
@@ -3,5 +3,5 @@ package com.softwaremill.macwire
 import cats.effect._
 
 package object autocats {
-    def autowire[T](dependencies: Any*): Resource[IO, T] = macro MacwireCatsEffectMacros.autowire2_impl[T]
+    def autowire[T](dependencies: Any*): Resource[IO, T] = macro MacwireCatsEffectMacros.autowire_impl[T]
 }
